feat(benchmarks): provision StarlingMonkey + ComponentizeJS lane with Wizer + Weval (#1125)
The harness already supported a STARLINGMONKEY_ADAPTER hook, but no adapter
was provisioned and the lane reported `unavailable` by default. This change
ships a reproducible adapter built around `@bytecodealliance/componentize-js`
with Wizer pre-init (always on) and Weval AOT specialization (default-on),
so the lane reports `ok` instead of `unavailable` whenever ComponentizeJS is
installed and `wasmtime` is on PATH.

- benchmarks/competitive/sm-componentize-adapter.mjs (new): canonical
  adapter location per the issue spec. Defaults Weval AOT to on; Wizer is
  always on (built into ComponentizeJS). Exposes explicit
  `wizerEnabled` / `wevalAotEnabled` in the sidecar metadata.
- scripts/starlingmonkey-componentize-adapter.mjs: replaced with a thin
  shim that forwards to the canonical adapter, so older docs/env configs
  pointing at the old path keep working.
- benchmarks/compare-runtimes.ts: auto-detect the bundled adapter via
  `import.meta.resolve("@bytecodealliance/componentize-js")` so users
  don't have to set STARLINGMONKEY_ADAPTER manually. Renamed the lane
  label to `... (Wizer + Weval) -> Wasmtime` so it's obviously distinct
  from the runtime-eval lane. Surface Wizer/Weval status in lane notes.
- benchmarks/competitive/README.md: rewrote the section with required
  tools, env-var table, side-by-side comparison vs runtime-eval, caveats.
- scripts/setup-benchmark-vendors.mjs: synced inline README.

End-to-end probe: adapter produces a 14.8 MB Wizer+Weval-specialized
component for `array-sum.js` in ~4.4s. `tsc --noEmit` clean. Lane no
longer reports `STARLINGMONKEY_ADAPTER is not configured`.

CHECKLIST-FOXTROT
